+++ CHANGELOG.md
@@ -39,6 +39,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ### Features
 * (server/grpc) [\#1018](https://github.com/Finschia/finschia-sdk/pull/1018) support custom r/w gRPC options (backport #1017)
+* (x/auth) [\#1026](https://github.com/Finschia/finschia-sdk/pull/1026) add the api for querying next account number (backport #1011)
 
 ### Improvements